fix: Fixed relation between FareModel and ParkingModel and Mappers/Services flow. refactor: Split Fare and Parking into separate POST and PUT Dtos. test: Removed InitialMigration for further testing.
**WARN: CalculateTotalPrice(ParkingModel) not working as intended. Although values and logic seem correct as per debugging, it always returns 0.**

todo: Fix CalculateTotalPrice(ParkingModel) functionality.

+++ ParkingLotAPI/Data/Constants/SizeFareMods.cs
@@ -9,11 +9,11 @@
 			Large = 3
 		}
 
-		public static IReadOnlyDictionary<VehicleSize, double> VehicleSizeFareMods { get; } = new Dictionary<VehicleSize, double>()
+		public static IReadOnlyDictionary<VehicleSize, decimal> VehicleSizeFareMods { get; } = new Dictionary<VehicleSize, decimal>()
 		{
-			{ VehicleSize.Small, 1.00 },
-			{ VehicleSize.Medium, 1.50 },
-			{ VehicleSize.Large, 2.00 }
+			{ VehicleSize.Small, 1.00m },
+			{ VehicleSize.Medium, 1.50m },
+			{ VehicleSize.Large, 2.00m }
 		};
 	}
 }
